# Optimize bottlenecks