These are some bottlenecks that were causing significant performance overhead for me. The changes are rough, but if they look good overall I can clean them up