@yonaskolb 2020-03-26
Thanks for your work here @michaeleisel. Would you be able to post some timings before and after this PR (after a rebase) for a large project, and also for the performance tests. Thanks!