@michaeleisel 2020-03-26
For carthage dependencies, on my 8 (physical) core machine it reduces the time from 600ms to 200ms for a large project. For the JSON parsing, it goes from 1,500 ms to 300ms. for performance testing, i can take a look. but first, are we good with these at a high level? (i.e., excluding stuff like try vs. try!, which can easily be fixed)