@brentleyjones 2020-03-26
I can't speak for @yonaskolb, but I'm interested in this. From a cursory glance I'm good with it at a high level.