@michaeleisel 2020-03-26
ok i'm happy with it now personally as well. here are the perf test results: https://gist.github.com/michaeleisel/3d1a5316d233571ca5e2ec9531351c82