@brentleyjones 2020-04-03
More details on our performance bottleneck:

<img width="1192" alt="image" src="https://user-images.githubusercontent.com/158658/78368156-cd626080-7588-11ea-8709-7c9ca1f60cf5.png">

Seems it's all in project generation and writing. In particular these two functions are hot spots: https://github.com/yonaskolb/XcodeGen/blob/01c6959f487fd67f57d05f1fe82d1263771cda25/Sources/XcodeGenKit/SourceGenerator.swift#L80-L82
(or just  https://github.com/yonaskolb/XcodeGen/blob/master/Sources/XcodeGenKit/PBXProjGenerator.swift#L222)

and

https://github.com/tuist/XcodeProj/blob/master/Sources/XcodeProj/Utils/ReferenceGenerator.swift#L27