@michaeleisel 2020-04-03
Yeah, I've only ensured it improves my own use case. I spent some time with that `getAllSourceFiles`. Without a deep understanding, it seems somewhat difficult to optimize. Maybe we could open another issue about it.